feat(auth): implement plugin action audit event handling and geo context resolution
